Add explicit template instantiation for specific grid types
