+++ openvdb/openvdb/unittest/TestGrid.cc
@@ -13,6 +13,8 @@
 #include <iostream>
 #include <memory> // for std::make_unique
 
+#include "utilCustomTree.h" // allow testing of non-standard tree configurations
+
 #define ASSERT_DOUBLES_EXACTLY_EQUAL(expected, actual) \
     CPPUNIT_ASSERT_DOUBLES_EQUAL((expected), (actual), /*tolerance=*/0.0);
