+++ openvdb/openvdb/unittest/TestTreeIterators.cc
@@ -7,6 +7,7 @@
 #include <openvdb/Types.h>
 #include <openvdb/tools/LevelSetSphere.h> // for tools::createLevelSetSphere()
 
+#include "utilCustomTree.h" // allow testing of non-standard tree configurations
 
 #define ASSERT_DOUBLES_EXACTLY_EQUAL(expected, actual) \
     CPPUNIT_ASSERT_DOUBLES_EQUAL((expected), (actual), /*tolerance=*/0.0)
