+++ openvdb/openvdb/unittest/TestValueAccessor.cc
@@ -7,6 +7,8 @@
 #include <openvdb/tools/Prune.h>
 #include <type_traits>
 
+#include "utilCustomTree.h" // allow testing of non-standard tree configurations
+
 #define ASSERT_DOUBLES_EXACTLY_EQUAL(expected, actual) \
     CPPUNIT_ASSERT_DOUBLES_EQUAL((expected), (actual), /*tolerance=*/0.0);
