+++ openvdb/openvdb/unittest/utilCustomTree.h
@@ -1,0 +1,16 @@
+// Copyright Contributors to the OpenVDB Project
+// SPDX-License-Identifier: MPL-2.0
+
+#ifndef OPENVDB_UNITTEST_UTIL_CUSTOM_TREE_HAS_BEEN_INCLUDED
+#define OPENVDB_UNITTEST_UTIL_CUSTOM_TREE_HAS_BEEN_INCLUDED
+
+// include tree definitions to allow for testing custom tree configurations
+
+#include "../tree/TreeImpl.h"
+#include "../tree/RootNodeImpl.h"
+#include "../tree/InternalNodeImpl.h"
+#include "../tree/LeafNodeImpl.h"
+#include "../tools/PointIndexGridImpl.h"
+#include "../points/PointDataGridImpl.h"
+
+#endif // OPENVDB_UNITTEST_UTIL_CUSTOM_TREE_HAS_BEEN_INCLUDED
